Start using XML API
